adding display measurements

+++ pressure/lps25hb.h
@@ -26,8 +26,8 @@
 uint8_t lps25hb_init(void);
 uint8_t lps25hb_read_byte(uint8_t reg_addr);
 void lps25hb_write_byte(uint8_t reg_addr, uint8_t value);
-int8_t lps25hb_get_temp();
-int16_t lps25hb_get_press();
+float lps25hb_get_temp();
+float lps25hb_get_press();
